move lazy array stuff to new arrays module
