+++ mesmerize_core/arrays/__init__.py
@@ -1,0 +1,7 @@
+from ._cnmf import LazyArrayRCM, LazyArrayRCB, LazyArrayResiduals
+
+__all__ = [
+    "LazyArrayRCM",
+    "LazyArrayRCB",
+    "LazyArrayResiduals"
+]
